LayoutTests/platform/gtk-wk1/fast/loader/show-only-one-beforeunload-dialog-expected.txt

CONFIRM: This is beforeunload from the top level frame.
CONSOLE MESSAGE: Blocked attempt to show multiple beforeunload confirmation dialogs for the same navigation.
CONSOLE MESSAGE: Blocked attempt to show multiple beforeunload confirmation dialogs for the same navigation.
